move search.highest_version() to util
